Move Size2i to net.minecraftforge.common.util
